+++ commands/convert.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"fmt"
-	"os"
 	"scan2epub/service"
 	"scan2epub/utils"
 
@@ -14,31 +13,20 @@ var ConvertCommand = cli.Command{
 	Usage:     "Convert scan to epub",
 	ArgsUsage: "<chap> [chap...]",
 	Aliases:   []string{"c"},
-	Flags:     convertFlags,
 	Action:    convertAction,
 }
 
-var convertFlags = []cli.Flag{
-	&cli.StringFlag{
-		Name:    "output",
-		Aliases: []string{"o"},
-		Usage:   "output directory",
-		Action: func(c *cli.Context, value string) error {
-			if value != "" {
-				os.Unsetenv("EPUB_DIR")
-				os.Setenv("EPUB_DIR", value)
-			}
-			return nil
-		},
-	},
-}
-
 func convertAction(c *cli.Context) error {
+	log, err := utils.GetLog()
+	if err != nil {
+		return err
+	}
+
 	if c.NArg() == 0 {
 		return fmt.Errorf("no chapter specified")
 	}
 
-	fmt.Printf("Converting %d chapters\n", c.NArg())
+	log.Printf("Converting %d chapters\n", c.NArg())
 	chaps := c.Args().Slice()
 	if err := service.Scan2Epub(chaps); err != nil {
 		return err